debug hex

+++ csf_assign01/fixedpoint.c
@@ -94,7 +94,7 @@ Fixedpoint fixedpoint_create_from_hex(const char *hex) { // Hex to decimal
   // Whole part
   // Test value: char whole_arr[2] = "2b";
   int length = whole_ctr;
-  int whole_sum = 0;
+  uint64_t whole_sum = 0;
   uint64_t base = 1;
   // printf("\n%d\n",length);
   for(int i = length-1; i>=0; i--) { // Traverse from end, sixeof
@@ -112,7 +112,7 @@ Fixedpoint fixedpoint_create_from_hex(const char *hex) { // Hex to decimal
   // printf("%lu",whole_sum);
   fp.whole = whole_sum;
   // Fractional part
-  int frac_sum = 0;
+  uint64_t frac_sum = 0;
   base = pow(16,15);
   printf("\n%lu\n",base);
   printf("BEFORE LOOP\n");
